+++ ecommerce/src/main/java/com/baron/ecommerce/repository/ItemsRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ItemsRepository extends JpaRepository<Items, Long> {
+public interface ItemsRepository extends JpaRepository<Items, Integer> {
 
     List<Items> findItemsByCartIdAndStatus(int cart_id, ItemStatus status);
 }
